move new BranchOffset16 to regmach module
